requirements.txt
certifi
charset-normalizer
et_xmlfile
idna
Levenshtein
numpy
openpyxl
pandas
psycopg2-binary
python-dateutil
python-dotenv
pytz
requests
six
tzdata
urllib3
apache-airflow
